Use vm module instead of `eval` in TransformData stream
